Move Jest files to own folder
