+++ setupTests.js
@@ -1,5 +1,5 @@
 import './src/global';
-import './tests/throwPropTypeErrors';
+import './tests/jest/throwPropTypeErrors';
 import { configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
