+++ tests/jest/throwPropTypeErrors.js
